#35 :: Refactor the previous successful hack
